requirements.txt
python-wordpress-xmlrpc==2.3
dataset==0.5.4
